requirement.txt
pytz
pymongo
requests
discord.py[voice]>=1.4.1
